feat: add package service
